+++ example/lib/main.dart
@@ -1,7 +1,8 @@
-import 'package:flutter/material.dart';
 import 'package:flutter_powerkit/flutter_powerkit.dart';
+// import 'package:flutter_powerkit/flutter_powerkit.dart';
 
-void main() {
+void main() async {
+  await UtilityHelper.instance.initiate();
   runApp(const MainApp());
 }
 
@@ -10,15 +11,6 @@ class MainApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        body: Center(
-          child: Text(
-            'Hello World!2'.capitalize(),
-            style: context.titleMedium,
-          ),
-        ),
-      ),
-    );
+    return const MaterialApp(home: FlutterPowerKitSplash());
   }
 }